Breaking out the controller functionality into it's own module.
This will help if you have to use multiple classes to inherit from.
Using your own controller you can now just include the module.

    class MyController < OtherBaseController
      include JkoApi::Controller
    end
