+++ lib/jko_api.rb
@@ -5,6 +5,7 @@ require "responders"
 require "jko_api/version"
 require "jko_api/class_descendants_builder"
 require "jko_api/constraints"
+require "jko_api/controller"
 require "jko_api/controller_helpers"
 require "jko_api/middleware"
 require "jko_api/request_error"
